% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCSAnnData.R
\name{GCSAnnData}
\alias{GCSAnnData}
\title{R6 Class for GCSAnnData}
\description{
R6 Class for GCSAnnData

R6 Class for GCSAnnData
}
\details{
An R6 class that wraps the Python GCSAnnData object for accessing h5ad files
stored on Google Cloud Storage.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{py_object}}{The underlying Python GCSAnnData object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GCSAnnData-new}{\code{GCSAnnData$new()}}
\item \href{#method-GCSAnnData-shape}{\code{GCSAnnData$shape()}}
\item \href{#method-GCSAnnData-var_names}{\code{GCSAnnData$var_names()}}
\item \href{#method-GCSAnnData-obs_names}{\code{GCSAnnData$obs_names()}}
\item \href{#method-GCSAnnData-get_rows}{\code{GCSAnnData$get_rows()}}
\item \href{#method-GCSAnnData-get_columns}{\code{GCSAnnData$get_columns()}}
\item \href{#method-GCSAnnData-print}{\code{GCSAnnData$print()}}
\item \href{#method-GCSAnnData-clone}{\code{GCSAnnData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GCSAnnData-new"></a>}}
\if{latex}{\out{\hypertarget{method-GCSAnnData-new}{}}}
\subsection{Method \code{new()}}{
Create a new GCSAnnData object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GCSAnnData$new(gcs_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gcs_path}}{Path to the h5ad file in GCS (gs://bucket-name/path/to/file.h5ad)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new GCSAnnData object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GCSAnnData-shape"></a>}}
\if{latex}{\out{\hypertarget{method-GCSAnnData-shape}{}}}
\subsection{Method \code{shape()}}{
Get shape of the AnnData object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GCSAnnData$shape()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector of (n_obs, n_vars)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GCSAnnData-var_names"></a>}}
\if{latex}{\out{\hypertarget{method-GCSAnnData-var_names}{}}}
\subsection{Method \code{var_names()}}{
Get column names (var_names)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GCSAnnData$var_names()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector of variable names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GCSAnnData-obs_names"></a>}}
\if{latex}{\out{\hypertarget{method-GCSAnnData-obs_names}{}}}
\subsection{Method \code{obs_names()}}{
Get row names (obs_names)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GCSAnnData$obs_names()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector of observation names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GCSAnnData-get_rows"></a>}}
\if{latex}{\out{\hypertarget{method-GCSAnnData-get_rows}{}}}
\subsection{Method \code{get_rows()}}{
Get rows from the X matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GCSAnnData$get_rows(
  indices,
  as_df = FALSE,
  add_names = FALSE,
  zero_based = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{indices}}{Vector of row indices (1-based) or row names}

\item{\code{as_df}}{Logical indicating whether to return a data.frame (default: FALSE)}

\item{\code{add_names}}{Logical indicating whether to add row and column names (default: TRUE)}

\item{\code{zero_based}}{Logical indicating whether indices are 0-based (default: FALSE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A sparse Matrix object or data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GCSAnnData-get_columns"></a>}}
\if{latex}{\out{\hypertarget{method-GCSAnnData-get_columns}{}}}
\subsection{Method \code{get_columns()}}{
Get columns from the X matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GCSAnnData$get_columns(
  indices,
  as_df = FALSE,
  add_names = FALSE,
  zero_based = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{indices}}{Vector of column indices (1-based) or column names}

\item{\code{as_df}}{Logical indicating whether to return a data.frame (default: FALSE)}

\item{\code{add_names}}{Logical indicating whether to add row and column names (default: TRUE)}

\item{\code{zero_based}}{Logical indicating whether indices are 0-based (default: FALSE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A sparse Matrix object or data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GCSAnnData-print"></a>}}
\if{latex}{\out{\hypertarget{method-GCSAnnData-print}{}}}
\subsection{Method \code{print()}}{
Print information about the GCSAnnData object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GCSAnnData$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GCSAnnData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GCSAnnData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GCSAnnData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
