% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcsAnndataR-package.R
\docType{package}
\name{gcsAnndataR}
\alias{gcsAnndataR}
\title{gcsAnndataR: Access h5ad files on Google Cloud Storage from R}
\description{
This package provides R wrappers for the gcs_anndata Python package,
allowing R users to access h5ad files stored on Google Cloud Storage
without downloading the entire file.
}
