% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{import_gcs_anndata}
\alias{import_gcs_anndata}
\title{Import the gcs_anndata Python module}
\usage{
import_gcs_anndata()
}
\value{
The imported Python module
}
\description{
Import the gcs_anndata Python module
}
\keyword{internal}
