% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_gcs_anndata}
\alias{install_gcs_anndata}
\title{Install the gcs_anndata Python package}
\usage{
install_gcs_anndata(method = "auto", conda_env = "r-reticulate")
}
\arguments{
\item{method}{Installation method to use}

\item{conda_env}{Conda environment to use}
}
\description{
This function installs the gcs_anndata Python package using reticulate.
}
