% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_gcs_h5ad}
\alias{read_gcs_h5ad}
\title{Read an h5ad file from Google Cloud Storage}
\usage{
read_gcs_h5ad(gcs_path)
}
\arguments{
\item{gcs_path}{Path to the h5ad file in GCS (gs://bucket-name/path/to/file.h5ad)}
}
\value{
A GCSAnnData object
}
\description{
Read an h5ad file from Google Cloud Storage
}
