% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_h5ad.R
\name{read_h5ad}
\alias{read_h5ad}
\title{Read H5AD}
\usage{
read_h5ad(path, to = "GCSAnnData", mode = "r", ...)
}
\arguments{
\item{path}{Path to the H5AD file to read. If the path starts with "gs://",
it will be treated as a Google Cloud Storage path and will automatically use
the GCSAnnData backend.}

\item{to}{The type of object to return. Must be one of: "InMemoryAnnData",
"HDF5AnnData", "GCSAnnData", "SingleCellExperiment", "Seurat". Default is "GCSAnnData".
Note that for paths starting with "gs://", this parameter is ignored and "GCSAnnData" is used.
A warning will be issued if a different value is provided for GCS paths.}

\item{mode}{The mode to open the HDF5 file. Default is "r" (read-only).
  Note that for paths starting with "gs://", this parameter is ignored and read-only mode is used.
  A warning will be issued if a different value is provided for GCS paths.

  * `a` creates a new file or opens an existing one for read/write.
  * `r` opens an existing file for reading.
  * `r+` opens an existing file for read/write.
  * `w` creates a file, truncating any existing ones.
  * `w-`/`x` are synonyms, creating a file and failing if it already exists.}

\item{...}{Extra arguments provided to `adata$to_SingleCellExperiment()` or
`adata$to_Seurat()`. See [AnnData()] for more information on the arguments of
these functions. Note: update this documentation when
[`r-lib/roxygen2#955`](https://github.com/r-lib/roxygen2/issues/955) is resolved.}
}
\value{
The object specified by `to`
}
\description{
Read data from a H5AD file, either from a local file or from Google Cloud Storage
}
\examples{
h5ad_file <- system.file("extdata", "example.h5ad", package = "anndataR")

# Read the H5AD as a SingleCellExperiment object
if (requireNamespace("SingleCellExperiment", quietly = TRUE)) {
  sce <- read_h5ad(h5ad_file, to = "SingleCellExperiment")
}

# Read the H5AD as a Seurat object
if (requireNamespace("SeuratObject", quietly = TRUE)) {
  seurat <- read_h5ad(h5ad_file, to = "Seurat")
}

# Read an H5AD file from Google Cloud Storage
\dontrun{
# This will automatically use GCSAnnData
gcs_adata <- read_h5ad("gs://my-bucket/my-file.h5ad")

# Access specific rows and columns
gcs_adata$X[1:10, 1:5]

# Access using names
gcs_adata$X[c("cell1", "cell2"), c("gene1", "gene2")]
}
}
