% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_gcs_anndata}
\alias{install_gcs_anndata}
\title{Install the gcs_anndata Python package}
\usage{
install_gcs_anndata(...)
}
\arguments{
\item{...}{Additional arguments passed to reticulate::py_install()}
}
\value{
Invisible NULL, called for side effect of installing the package
}
\description{
This function installs the gcs_anndata Python package using reticulate.
The gcs_anndata package allows for reading AnnData objects directly from
Google Cloud Storage.
}
\examples{
\dontrun{
install_gcs_anndata()
}

}
\seealso{
\code{\link[reticulate]{py_install}} for more details on installation options
}
